refactor: extract backend trending podcast mock handler to separate file
